version finale, a la norme, normalement les include sont bonnes

+++ get_next_line.h
@@ -6,17 +6,18 @@
 /*   By: groubaud <groubaud@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/13 10:13:59 by groubaud          #+#    #+#             */
-/*   Updated: 2021/06/20 10:48:29 by groubaud         ###   ########.fr       */
+/*   Updated: 2021/06/21 14:53:48 by groubaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef GET_NEXT_LINE_H
 # define GET_NEXT_LINE_H
 
+# include <unistd.h>
 # include <stdlib.h>
 
 char	*ft_strjoin_gnl(char *s1, char const *s2);
-int		ft_isinstr(char c, const char *str);
+ssize_t	ft_isinstr(char c, const char *str);
 size_t	ft_strlen(const char *str);
 char	*ft_strdup(const char *s);
