+++ get_next_line_utils_bonus.c
@@ -1,19 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   get_next_line_utils.c                              :+:      :+:    :+:   */
+/*   get_next_line_utils_bonus.c                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: groubaud <groubaud@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/13 10:13:42 by groubaud          #+#    #+#             */
-/*   Updated: 2021/06/20 10:44:19 by groubaud         ###   ########.fr       */
+/*   Updated: 2021/06/21 14:44:46 by groubaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <sys/types.h>
 #include <stdlib.h>
 
-int	ft_isinstr(char c, const char *str)
+ssize_t	ft_isinstr(char c, const char *str)
 {
 	ssize_t		index;
